# openApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ImmichFrame.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Admin/Setup": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "SetupAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSetupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSetupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSetupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsUpdateResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsUpdateResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsUpdateResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Admin/Settings": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "GetAdminSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSettings"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSettings"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "UpdateAdminSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerSettings"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerSettings"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServerSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsUpdateResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsUpdateResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsUpdateResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Admin/Settings/TestAccount": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "TestAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerAccountSettings"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerAccountSettings"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServerAccountSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTestResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTestResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTestResultDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountTestResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AdminSetupDto": {
        "type": "object",
        "properties": {
          "adminPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GeneralSettings": {
        "type": "object",
        "properties": {
          "downloadImages": {
            "type": "boolean"
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "imageLocationFormat": {
            "type": "string",
            "nullable": true
          },
          "photoDateFormat": {
            "type": "string",
            "nullable": true
          },
          "interval": {
            "type": "integer",
            "format": "int32"
          },
          "transitionDuration": {
            "type": "number",
            "format": "double"
          },
          "showClock": {
            "type": "boolean"
          },
          "clockFormat": {
            "type": "string",
            "nullable": true
          },
          "clockDateFormat": {
            "type": "string",
            "nullable": true
          },
          "showProgressBar": {
            "type": "boolean"
          },
          "showPhotoDate": {
            "type": "boolean"
          },
          "showImageDesc": {
            "type": "boolean"
          },
          "showPeopleDesc": {
            "type": "boolean"
          },
          "showTagsDesc": {
            "type": "boolean"
          },
          "showAlbumName": {
            "type": "boolean"
          },
          "showImageLocation": {
            "type": "boolean"
          },
          "primaryColor": {
            "type": "string",
            "nullable": true
          },
          "secondaryColor": {
            "type": "string",
            "nullable": true
          },
          "style": {
            "type": "string",
            "nullable": true
          },
          "baseFontSize": {
            "type": "string",
            "nullable": true
          },
          "showWeatherDescription": {
            "type": "boolean"
          },
          "weatherIconUrl": {
            "type": "string",
            "nullable": true
          },
          "imageZoom": {
            "type": "boolean"
          },
          "imagePan": {
            "type": "boolean"
          },
          "imageFill": {
            "type": "boolean"
          },
          "playAudio": {
            "type": "boolean"
          },
          "layout": {
            "type": "string",
            "nullable": true
          },
          "renewImagesDuration": {
            "type": "integer",
            "format": "int32"
          },
          "webcalendars": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "refreshAlbumPeopleInterval": {
            "type": "integer",
            "format": "int32"
          },
          "weatherApiKey": {
            "type": "string",
            "nullable": true
          },
          "unitSystem": {
            "type": "string",
            "nullable": true
          },
          "weatherLatLong": {
            "type": "string",
            "nullable": true
          },
          "webhook": {
            "type": "string",
            "nullable": true
          },
          "authenticationSecret": {
            "type": "string",
            "nullable": true
          },
          "adminPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ServerAccountSettings": {
        "type": "object",
        "properties": {
          "immichServerUrl": {
            "type": "string",
            "nullable": true
          },
          "apiKey": {
            "type": "string",
            "nullable": true
          },
          "apiKeyFile": {
            "type": "string",
            "nullable": true
          },
          "showMemories": {
            "type": "boolean"
          },
          "showFavorites": {
            "type": "boolean"
          },
          "showArchived": {
            "type": "boolean"
          },
          "showVideos": {
            "type": "boolean"
          },
          "imagesFromDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "imagesFromDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "imagesUntilDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "albums": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "excludedAlbums": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "people": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ServerSettings": {
        "type": "object",
        "properties": {
          "General": {
            "$ref": "#/components/schemas/GeneralSettings"
          },
          "Accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerAccountSettings"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SettingsUpdateResultDto": {
        "type": "object",
        "properties": {
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}